Remove excess component file and render contents within parent
